Hotfix for failed load? Disabled db housekeeping in startup. Crash not reproducable in local environment
